requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.3
certifi==2020.6.20
chardet==3.0.4
click== 8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
SQLAlchemy==1.4.39
email-validator==1.2.1
urllib3==1.25.10
visitor==0.1.3
Werkzeug==2.1.2
WTForms== 3.0.1
python-dotenv==0.20.0
